+++ packages/fluentui/docs/src/examples/components/Datepicker/index.tsx
@@ -2,11 +2,13 @@ import * as React from 'react';
 
 import Types from './Types';
 import States from './States';
+import Usage from './Usage';
 
 const DatepickerExamples = () => (
   <>
     <Types />
     <States />
+    <Usage />
   </>
 );