+++ packages/fluentui/react-northstar/src/index.ts
@@ -142,6 +142,8 @@ export * from './components/Skeleton/Skeleton';
 export * from './components/Skeleton/SkeletonLine';
 export * from './components/Skeleton/SkeletonShape';
 export * from './components/Skeleton/SkeletonButton';
+export * from './components/Skeleton/SkeletonText';
+export * from './components/Skeleton/SkeletonInput';
 export * from './components/Skeleton/SkeletonAvatar';
 
 export * from './components/Slider/Slider';