+++ packages/fluentui/react-northstar/src/themes/teams/componentVariables.ts
@@ -91,6 +91,8 @@ export { skeletonVariables as Skeleton } from './components/Skeleton/skeletonVar
 export { skeletonVariables as SkeletonLine } from './components/Skeleton/skeletonLineVariables';
 export { skeletonVariables as SkeletonShape } from './components/Skeleton/skeletonShapeVariables';
 export { skeletonVariables as SkeletonButton } from './components/Skeleton/skeletonButtonVariables';
+export { skeletonVariables as SkeletonText } from './components/Skeleton/skeletonTextVariables';
+export { skeletonVariables as SkeletonInput } from './components/Skeleton/skeletonInputVariables';
 export { skeletonVariables as SkeletonAvatar } from './components/Skeleton/skeletonVariables';
 
 export { sliderVariables as Slider } from './components/Slider/sliderVariables';